refactor: reuse getNextPageParam logic across queries
