+++ src/api/purchase/get-purchases.ts
@@ -1,6 +1,6 @@
 import { createInfiniteQuery } from 'react-query-kit';
 
-import { client } from '../common';
+import { client, DEFAULT_PAGE_PARAMS } from '../common';
 import { API_CONSTS, QUERY_KEYS } from '../consts';
 import type { PurchasesResponse } from './types';
 
@@ -22,12 +22,5 @@ export const usePurchases = createInfiniteQuery({
     });
     return data;
   },
-  getNextPageParam: (lastPage) => {
-    const nextUrl = lastPage.pagination.next_url;
-    const queryString = nextUrl.split('?')[1];
-    const urlParams = new URLSearchParams(queryString);
-    const page = urlParams.get('page');
-    return page ? parseInt(page, 10) : undefined;
-  },
-  initialPageParam: 1,
+  ...DEFAULT_PAGE_PARAMS,
 });
